requirements.txt
streamlit==1.24.0
langchain==0.0.300
langchain-huggingface==0.0.3
langchain-pinecone==0.0.3
pinecone==5.0.1
sentence-transformers==2.2.2
requests==2.31.0
pillow==10.0.0
python-dotenv==1.0.0
watchdog==4.0.0